requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.7
django-bootstrap5==23.3
django-cpf==0.1.0
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
django-wkhtmltopdf==3.4.0
environs[django]==9.5.0
gunicorn==21.2.0
marshmallow==3.20.1
packaging==23.2
phonenumbers==8.13.24
phonenumberslite==8.13.24
psycopg[binary]==3.1.13
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
whitenoise==6.6.0
